Simplify storage class
