+++ src/app/backend/resource/storageclass/common.go
@@ -15,31 +15,9 @@
 package storageclass
 
 import (
-	"github.com/kubernetes/dashboard/src/app/backend/api"
 	"github.com/kubernetes/dashboard/src/app/backend/resource/dataselect"
-	"github.com/kubernetes/dashboard/src/app/backend/resource/persistentvolume"
 	storage "k8s.io/api/storage/v1"
 )
-
-func toStorageClass(storageClass *storage.StorageClass) StorageClass {
-	return StorageClass{
-		ObjectMeta:  api.NewObjectMeta(storageClass.ObjectMeta),
-		TypeMeta:    api.NewTypeMeta(api.ResourceKindStorageClass),
-		Provisioner: storageClass.Provisioner,
-		Parameters:  storageClass.Parameters,
-	}
-}
-
-func toStorageClassDetail(storageClass *storage.StorageClass,
-	persistentVolumeList *persistentvolume.PersistentVolumeList) StorageClassDetail {
-	return StorageClassDetail{
-		ObjectMeta:           api.NewObjectMeta(storageClass.ObjectMeta),
-		TypeMeta:             api.NewTypeMeta(api.ResourceKindStorageClass),
-		Provisioner:          storageClass.Provisioner,
-		Parameters:           storageClass.Parameters,
-		PersistentVolumeList: *persistentVolumeList,
-	}
-}
 
 // The code below allows to perform complex data section on []storage.StorageClass
